Shopping list: change a misleading message.
" and buy it" suggests it's an immediate purchase, while in fact it's the
only way to remove the item from the shopping list.  When removed, it becomes
selected for purchase, but you're in no way forced to continue with it.
(cherry picked from commit c43b97f86c25d6a02e4dea7557095cfff2a441d9)