Don't hatch/teleport into cabins on the troll bridge.
On D:9, rock trolls are massively out of depth, they are killable only with
appropriate strategy/kiting -- but are not too deadly because of their
slowness.  On the other hand, being stuck with two rock trolls and one
vanilla is something no one can survive.

No idea how a free space could have generated with the secret door still
unknown, but at least once it did.