Make Nemelex's handling of corpses more consistent.
Nemelex used to compute corpse value using three different formulas: for timeout
decrease, piety gain, and deck weighting.

Now Nemelex uses the same formula for all three, and it is a simple one: every
corpse has value 50.

The effects of this:
1) Piety gain from corpses is very roughly the same on average, but it decreases
much more smoothly as piety increases.
2) Corpses have a far smaller effect on deck weighting now, since previously
we used value = corpse mass (times 10). Decks of summonings tended to dominate
the weightings, so this is hopefully a change for the better, but it will
require testing. Also, the fact that corpse mass mattered was inconsistent
with sacrificing corpses to other gods.