Ziggurat themed sprint map. (st)
https://crawl.develz.org/mantis/view.php?id=5323

This sprint map has 27 rooms with randomized themed monster sets, and loot.
It's much like a ziggurat, but contained all in a single map, and with the
rooms linked by teleporters instead of stairs. You get the Orb after room
27.

All the items are pre-identified. Note that there is a problem here with
autoinscriptions not being given to pre-identified artefacts.

See the mantis issue for st's comments.