Drop APPEAR: for Wrath of Trog, Variability and Leech, id the Singing Sword on wield.
For the first three, the colour and tile gives them away anyway, even though
their names could appear on randarts.

For the Singing Sword, I'm leaving the old behaviour because of its greeting
gimmick, although it reveals the pluses on wield now.  It's iffy and
inconsistent with the rest -- ideas?