Finally fix capitalization of "the Shining One" in the code.
At the beginning of a sentence, it is "The Shining One", as expected,
but it's now lowercase elsewhere, as the god's name is now lowercase.
For simplicity, however, the database entries are still under the
capitalized form.

The code now puts god_name() inside uppercase_first() in the case of the
beginning of a sentence, generically so that any more properly
lowercase-named gods added later will be handled properly.

Note that simple_god_message() automatically does the capitalization fix
(as it already fixes apostrophes).  Ideally, god names should always be
accessed via god_name() instead of being hardcoded, but that's a much
more involved fix.