Let b0rsuk know what the hell a bardiche is.
These translations are here to let us have something to test the code on,
and to identify related problems.  Putting more effort right now is probably
a waste -- the format may change, translations in general may turn out to be
too hard to do (descs are a low hanging fruit), etc.