Declare an explicit type for colours.
An imperial buttload of places assumes these to be 8 bits, which makes it
impossible to add support for 256 (or more!) colours.  This commit tries
to allow expanding it in the future; sadly, it is badly incomplete.  At
least, new code can use the new type.

Also, beware of "branded" colours which use bits above 8 for inverse,
underline, etc.