Remove placeholder Spider ending and add a new one.
Mine could still use a bit of work - it needs to be worthy of being a branch
ending, but it's hard to get a feel for what the difficulty is until a few
people play it. Making a vault harder or easier is trivial, of course. I'm
thinking, the main hard enemies you want to use in a Spider ending are ghost
moths and emperor scorpions. _Maybe_ moths of wrath (I haven't used any
here). Obviously none of those are spiders, but there aren't really any
"boss" spider enemies at the moment.