Teleport away monsters unsubmerging on the player's square.
If an unsubmerging monster:
  (1) is on the same grid as the player (or a monster, though that
      shouldn't happen),
  (2) could not be moved to an adjacent submergable square,
  (3) cannot remain submerged on its current square, and
  (4) could not be moved to an adjacent habitable square;
then it can unsubmerge on top of the player, producing a "same square as
you" warning.  Usually this situation (in particular, (3)) occurs when
the monster's grid changed to no longer support its submersion, in which
case _dgn_check_terrain_player() would teleport the monster away.
However, it can also occur with air elementals' involuntary submersion,
without the grid having changed at all; in this case, nothing prevents
the monster from occupying the same space as the player, and things such
as the end of flight will trigger an assertion.

Instead, trigger the teleport whenever we would give the "same square as
you" warning; and only give that warning if the teleport failed.  The
teleport in _dgn_check_terrain_player() is therefore no longer needed.

Also, when trying to move the monster in _prepare_del_ench(), look at
squares adjacent to the monster, not to the player.  This could differ
if the monster was trying to unsubmerge on top of another monster rather
than the player (which, again, shouldn't happen, but . . .).

Fixes #4158, #5419.