Make walls from the Tomb card expire after a while, depending on power
The duration is randomised but with a minimum based on power. Even at
high power it might not be enough to rest to full health (loads of time
to use consumables to heal safely, though). Duration might need further
tweaking, not sure.

Also fixes Khufu's tomb removing nearby walls that weren't originally
part of the tomb when it disappears.