Move the cursor even for muted prompts.
Previously, when the prompt message was muted, the cursor would remain
in GOTO_DNGN; typing a long enough input would crash the game as the
cursor goes out of bounds.  Instead, show messages (and thus move the
cursor) if the prompt is muted.  You still don't see the prompt---
presumably you had it muted for a reason.

Fixes #5435.