Fix stat gain on interrupted level-up.
The commit "Redraw the screen while gaining a level" swapped the order
of the stat gain and level increase in order to display the new level at
the prompt.  However, the order was important: if a SIGHUP
is received at the stat-gain prompt and the player later reconnects,
it is not having the new experience level that causes the stat gain
prompt to be displayed again.

We leave in the redraw, even though it will now display the previous
experience level.

This partially reverts commit 844e08c9dd00e13cbd29468fd38713c5a87cf085.