New Nemelex ability: Deal Four.
You draw four cards from the top of the deck and play them immediately
(in sequence, but taking a single turn).  To balance this:

 * It doesn't work on marked decks.
 * The MP, piety, and hunger costs are quite high.
 * The rest of the deck is destroyed.
 * If there weren't enough cards, you "get to" deal a card from the deck
   of punishments (just one, not the number of cards you lacked).

This probably needs more tweaking.  In particular, the MP cost is high
(to limit spamming the ability in combat), but it may be so high as to
cause a problem for non-casters of certain races.