New monster: orb spider.
A spider with IOOD and the new M_MAINTAIN_RANGE flag.  Does not spawn
randomly currently; has a vault on Spider:4-5 (default weight) for
testing purposes.  Numbers probably need tuned, and it may need to be
toned down with more cantrip slots.

The tile is currently just a recolouring of the spider tile.  That,
and the text, could use some improvement.