A few vault tweaks.
-Corrected a few "autoexplore traps"
-Added loot to some highly dangerous vaults.

The reason for the latter is simply that extreme risk should have some
reward attached to it. Said reward doesn't have to be huge though.

There was talk in ##crawl-dev about also addressing some early vaults which
place OOD monsters. I've looked at these, and it seems better to simply
provide the player with warning (ie, make the vault obvious, and stop
explore before the player gets too far). An '8' monster on D:5 tends to be
dangerous, but not absurd if it's placed correctly.