Changes to Spider:5.
The vault is now placed along the edge. There are also a few doors inside
it. This should help with some web related issues KoboldLord pointed out.

I've also changed the monster set a bit. The orb spiders might not work so
well in an enclosed vault (they're meant for open areas). On the other hand,
the "loop" layout of this vault takes advantage of their special AI.