Slightly reduce redback's HP and increase wolf spider's.
This is only a small adjustment. Redbacks are SIZE_TINY, while wolf spiders are
SIZE_LARGE. It makes sense for the former to be a bit more frail and the
latter a bit more robust, compared to other spiders. HP values (or HD) for
other spiders might still be worth adjusting.