Don't crash when reading a line with mute:. and clear_messages .
If there were no messages at all, message_window::place_cursor() would
not move the cursor, leaving it in the last row and column.  This would
the cause line_reader::cursorto to try to scroll the messages, leading
to "ASSERT(next_line >= n) in 'message.cc' at line 449 failed."

This does not seem to happen if the cursor was legitimately in the last
row and column, so the crash appears to be avoided entirely.