Better randomization of shafts and hatches destinations.
The old code set the destination of hatches when they were created and if it was a wall or
another unhabitable feature, the player was moved to the closest habitable cell.
In dense levels like Orc, the player often ended up against a wall or in a corner.

The new code choose the random destination when entering the level. When a suitable destination
is found it is saved in a map marker located at the original position.
There is a little bug. If you take a hatch down on level n-1 and then find a hatch up on level n+1
at the same coordinates, then the 2 hatches will have the same destination in level n.
This is extremely rare, unabusable and unoticeable, so I wouldn't worry about it. Fixing it would
require changing map_position_marker to store the dungeon_feature_type.