Add explore_stop = portals, branches
The former was already implemented, but could not be specified
in the init file (and thus could not be turned off with -=).
Fix that, and also make Lab entrances count as portals.

The "branches" setting includes branch entrances as well as the
entrances to Hell, Abyss, and Pandemonium.  It does not include
portal vaults.

This is a regression for those who use "explore_stop -= stairs"
and expect it not to stop at branch entrances.