Fix weapons not corroding upon one-shotting a jelly.
Also, this commit makes attacker_sustain_passive_damage() called for
monsters as well.  The check for weapon corrosion applying to the player
only has been moved there for now, it'd be good to implement that for
everyone.  Not doing this outright as this would hurt Beoghites and thus
it may be better to discuss this first.