Fix a few more net/web messages.
Two cases: enchantment status text for monsters, and messages for caught
allies.  MB_CAUGHT is split into MB_CAUGHT and MB_WEBBED; and
held_status() is changed to take an optional actor pointer, defaulting
to &you.

The same tile flag is still used for both statuses, but it is now possible
to change this.