Suppression aura effects, part 1
This first round of suppression effects focuses solely on denying the player
any advantages (and disadvantages!) from equipped magical items. Specifically,
I searched for the following three strings in the source: "wearing",
"player_equip", and "scan_artefact" and added a "you.suppressed()" clause to
ensure that effects only apply outside of fields of magical suppression.
There's a lot more to do, but this is a good start. Specifically, this
addresses the most important aspect of the suppression aura, which is to deny
the player any resists from items.