Item effect -> player effect abstraction, part 1
There are a whole bunch of places in the code where effects on the player are determined by directly checking for whether a single item is equipped. For example, `wearing_amulet(AMU_STASIS)`. This commit abstracts all of these checks into actual effects, such as `player_effect_stasis()`, to more easily jive with checking for suppression. For some effects, especially resist mutation, more work still needs to be done to unify effect checking... but this suffices for the moment.