Use better hints for some hidden orc entries.
I just quit a game because I couldn't find one of these (mainly because I
wanted to find which vault was the culprit).

Apparently it used a metal wall as a hint, but that doesn't show up on the
map (in console) since everything is the same colour. Adding orcish idols
should help this.

This is practically a bug fix.