Replace Pit Fiends with Hell Sentinels
Pit Fiends were mainly just slower Brimstone Fiends. Hell Sentinels are
intended to be a a tough, highly resistant 1 to replace them, and to
hopefully still be a threat without having torment. Numbers and spells
and stuff are all open to improvement.

Thanks to st_ for the name and description.