Make moths web immune.
It's been commented several times now that it's unusual for ghost moths (a
key enemy in Spider) to be so vulnerable to webs.

Given that a new moth enemy is going to be added sometime soon, and moths in
general seem to be a sub-theme of the branch, I've gone and made the whole
genus web immune. All the other bugs (insects mostly) will stay vulnerable
to webs - they're allowed to spawn in the branch to give it a vague sense of
"ecosystem".

I'm aware that it's a little weird for all moths to be immune like this, but
I'm going to side with gameplay over realism on this. I guess Crawl's moths
have just evolved to be web immune, given their natural habitat is shared
with Spiders.