Remove a bogus and harmful assert about visits to Labyrinths.
You can visit the same Lab multiple times, due to banishment.

This whole code needs to be ripped away and rewritten, so I'm putting just
this stub in for 0.10.

Also, s/num_visits/levels_seen/ for Labs.  Portal vaults need a similar
treatment, but it's more complex there, and since 1. it's a cosmetic issue
(wrong count in morgues), and 2. see above, I'm skipping that.