Repaint the dungeon view after resizing.
This also fixes a crash, since the view buffer was in an undefined
state after the resize.

This doesn't deal with the text areas, though; also, there seems to be
a problem with the transforms somewhere which causes the whole window
content to be moved down when the window height gets smaller (but no
similar problem for the x axis, interestingly).