Use ♣ not Ω for trees in charset_ibm.
The only reason ♣ has been avoided before is that it's CP437 code, 5, is
interpreted as a control character by some terminals.  It had some point
when we used ncurses in 8-bit mode (where it relies on \e[11m).  The only
way to reproduce this problem now is to use a locale with CP437 encoding
-- not doable unless you write a locale definition yourself, something that
requires intent and knowledge.

On Windows, console output goes through Microsoft's non-standard API with
no support for control characters, so this not an issue there as well.