Slightly better default training values in auto mode.
Training is scaled the the square of the number of skill points. This means
that high level skills will start with a much higher training ratio than low
level ones.
It's not always the best strategy to train your highest level skill first
(priest) but it often is, and at the very least, it can be a hint to newbies
that they shouldn't spread their skill training too much in the begining.