Mark the Wild Magic mutation as "mostly bad".
With wild magic 2 (not a rare case with Xom), you get 68% failure even on
level 1 spells with D:6 skills, which means mostly -CAST.  Later on the
mutation may be more useful, but since at that point you're likely at the
power cap for low level spells anyway, it's still mostly bad.

Ie, it's in the same club as berserkitis and teleportitis, both of which
sometimes prove useful.