Use unseen background tile for (unknown) items.
If the item type is unknown but the item instance has known type (i.e.
an "(unknown)" object in a shop), the background tile for the item would
be the correct one for its appearance.  However, this is an information
leak: it allow the tiles player to ctrl-f an item in a shop to learn,
for example, that "potion of cure mutation (unknown)" has a blue tile,
and thus that this untested "silvery-white potion" must be something
else.

Fix the leak by using the base ("unseen") background tile for such
objects, along with the icon for the identified type.