Give the correct message for tabbing while webbed.
The clua you.caught() function now returns either a string (the result
of held_status(&you)) or nil.  It can still be used in a conditional
(since nil evaluates as false and all strings as true), but any client
code that uses something like "you.caught() == true" will need to be
changed.

Fixes #5552.