New option: fsim_scale.
It's used to configure which skills are used as a scale in
fsim_simple_scale mode. By default, only the weapon skill is scaled.

It takes a list of skills with an optional divider. weapon can be used for
the default weapon skill for the simulation. Setting this option to:
weapon, fighting/2
restores the behaviour of the old simulator.

if XL is specified instead of a skill, then it will be used as the scale.
Skills will be trained automatically, and the number provided for each
skill will be used to set the training. For example:
xl, weapon:2, fighting:1
means that 2/3 of xp will be used to train the weapon skill and 1/3 for
fighting.

/ and : are both recognised as separators in normal and xl mode. However,
given that there meaning is completely different, it's recommended to use
/ in normal mode and : in XL mode.