Remove the chance for easy monsters in lower D.
I'm not really sure what are going to be the consequences of this change.
At the very least, it should remove trivial monsters from the lower dungeon,
but it might also create more dangerous one and increase the global XP value
of the dungeon, so we should keep an eye on it.