Actually repaint after drawing the message window.
The call from redraw_screen() to display_message_window() was moved
after the call to viewwindow() some time ago, so that the cursor would
be placed correctly after a prompt.  However, this means the message
window is not actually flushed to the terminal (and the cursor is not
actually moved) until the next redraw.  This is most noticeable when
resizing the window: the message area will be blank until something
happens.