Ziggurats: fix wall colours in tiles.
Unlike console, where walls are constant for the whole zig, in tiles you get
wall_vault unless the monster set for the level contains the word "place",
in which case it is a random Zot wall.

I made it use the Zot wall that matches the console colour -- ie, constant
for the whole ziggurat.

Also, I really hate zig code operating on grd() directly rather than creating
a vault.  It'd make changing colours and tiles so much simpler, and allow
eliminating a number of functions not used elsewhere.