Trap unscaling, act 1: blade traps.
They're moved deeper (min. depth: D:12 -> D:15), and do the same damage
everywhere -- matching what they used to do on D:25.

The rationale is, it is spoiler knowledge that the same trap will suddenly
do twice the damage when found in a different place.  It can kill such
unspoiled players at no fault on their side.  Same applies to other traps,
and now mimics (as discussed on IRC recently).

I'm not sure if adjusting them towards the near max of their former damage
is ok, especially with moving the min depth by only 3.  Playing it safe on
the unsafe-for-players side for now.