Grey draconian changes.
Grey draconians were generally regarded as the weakest colour, and for good
reason - in exchange for a breath weapon and a resist, you just got better
Stealth/Stabbing apts (irrelevant in the vast majority of games and not
important enough to build a character around) and breathless
(very situational, mainly just used for traveling across water and self-meph).

Now they are themed more like iron dragons - still no breath weapon, and the
Stealth/Stabbing apts are normal now, but the scales are iron-like, giving
5 extra AC over other dracs. Also, the iron scales give them extra mass,
helping them to balance properly to attack in water without penalty.