Don't block distortion teleports of monsters (Patashu, G-Flex).
Commit 83dafbf made the stasis/-tele/sprint checks on distortion
teleports affect monsters too.  The first two don't make sense,
because it is the player's items being checked; and teleporting
monsters in Sprint is not a problem the way teleporting the player
is.

It would probably make sense for -TELE artefacts to have their effect on
monsters, but that would need checking in a lot more places than this,
remembering that monsters are sometimes teleported in "emergency"
situations such as being on the same square as the player, and those
teleports should not be blocked.  Such a change will have to wait for
another day.

Thanks to G-Flex for spotting this, and Patashu for tracking down the
erroneous code.