Remove rAcid from the % screen and give rCorr to yellow draconians.
rAcid still has 3 levels and rCorr amulet still stacks with cloak of
preservation to give 2 of them. It's just too rare to be worth a line in
the % screen.
Yellow scale mutation doesn't give rCorr and should probably be removed
for simplicity. It's one of the least relevant scale mutation anyway.