A shell script to optimize new tiles.
I was tired of looking up the commands in tiles_creation.txt every
time I add a tile, so I wrote a script that applies them to all png
files that are modified in the current index (so you have to git add
them first). I didn't test this much and don't have much experience in
shell scripting anyway, so maybe this can be improved, but it may be
useful to someone else.