Avoid info leak, get_items_at -> get_item_at.
Returns an item or nil.  This avoids leaking information about items
at the bottom of an unvisited pile.  For the same reason, do not return
anything from an invisible pile (e.g. one at the bottom of deep water
for non-swimmers).

It would be nice to use the stash to return an entire visited pile,
but there doesn't seem to be a way to get a semi-permanent pointer
to an item_def in the stash.