Return multiple items from items.get_items_at().
Like full_describe_view(), we get the extra items from the stash
tracker; on untracked levels we can get only the top.

Difficulties arise because the stash tracker returns item_defs on the
stack where they cannot safely be pointed to.  To handle this properly,
item_wrapper gains a new boolean member indicating whether the item_def
is a temporary copy (as opposed to an element of mitm, for example); a
new C function clua_push_item_temp() allocates a copy of an item_def and
creates such a wrapper.  If the temporary flag is true, the item_def
will be freed when the wrapper is GCed.