Improve KMASK/KMONS interaction, singlify Orb Guardians.
KMONS, KITEM, and KFEAT are intended to override the normal meaning of
the key glyph.  For example, with:
  MONS:  kobold
  KMONS: 1 = orc
the glyph '1' will produce an orc and not a kobold; and with:
  KMONS: W = orc
the glyph 'W' will produce an orc on floor, not shallow water.

KMASK works differently: it keeps the underlying meaning of the glyph.
This is accomplished by keyed_mapspec::set_mask() setting the feature
to the original glyph (commit 3ba8bc8).  So with:
  KMASK: W = no_monster_gen
the glyph 'W' will produce shallow water with the no_monster_gen mask,
and with:
  MONS:  kobold
  KMASK: 1 = no_trap_gen
the glyph '1' will produce a kobold on floor with the no_trap_gen mask.

This is all well and good.  However, when both KMONS (or KITEM or KFEAT)
and KMASK are specified, the glyph should be overridden.  However, since
set_mask() set the feature explicitly, this does not happen: both the
glyph and the KMONS have their effect.  That is:
  MONS:  kobold
  KMONS: 1 = orc
  KMASK: 1 = no_item_gen
results in the glyph '1' producing both a kobold and an orc.  The
difference in result from the first example is unacceptible.

Furthermore, a multi-glyph key in KMASK was applying the fix-up only for
the first glyph:
  KMASK: wW = no_pool_fixup
caused the 'W' glyph to become floor rather than shallow water.

Fix all of this by removing the special-case code from set_mask(), and
instead having _vault_grid() treat a mask-only mapspec the same as no
mapspec, using _vault_grid_glyph() rather than _vault_grid_mapspec().
The mask itself is handled elsewhere, in dgn_register_place().

Fixes 5561.

This reverts commit 3ba8bc84d018da1452880ac897a433ca7752ebb5.