Remove a few more gaps in DNGN_ enums.
MINWALK is referenced only in two places, in the unlikely case a new feature
you can fly but not walk over gets added, it'd be a matter of an || -- in fact,
the only non-dubious place with MINWALK already has two such cases.

There's nothing that relies on DNGN_EXPLORE_HORIZON not being saved, so we're
free to just add enums at the end.