Remove the zot5_orb subvault.
Replace it with kitem() and kmask() calls in the zot5_orb function.
This allows Zot:5 to be correctly regenerated with wizmode &^R:
Previously this did not work because the subvault was not freed
when wiping level, so it was no longer available when placing
the orb chamber a second time.