Set ZotDef in Zot.
Allow games to use different root branches.  Currently this is Zot for
zotdef, and Dungeon for everything else.

If there are reachable branches that precede the root branch in
branches[], the ^O screen will incorrectly print them first.  This is
not currently the case for any game type, and isn't likely to be the
case in the future, because branches[] is for the most part
topologically sorted.