Don't let 8/9 glyphs request OODs on a level without OOD data.
The assert is correct -- before portal_branches, requesting an OOD monster
resulted in disallowing any but most common ones, which happen to be on the
easy side (ie, contrary to what was asked for).

This fix is a kludge: really, we'd want to rethink the whole OOD thing,
especially the absdepth0 30 cap which makes little sense for branches other
than D, especially Zot and hells.