Fix the banishment area being misshapen if it's cut off by the map edge.
This caused a crash due to an off-by-one error that read an unitialized
place in memory (whatever was just past the array on the heap), which could
be an invalid feature.

It also made the new Abyss not resemble the place you came from.

This commit also converts shops to abandoned shops (instead of floor) and
portals to stone arches (rather than 1/3 arches 2/3 floor like stairs).