Make all missile brands visible.
Enchantment is still unknown.

Determining the brand was a matter of shooting a single shot, and then you
could keep a sample so other identical ammo would stack with it.  My recent
commit removed the need for the latter but only for fully identified
ammunition.

Note that 10 out of 16 randomly occuring brands already had this property
(half of them needle-only).  This was mostly contrary to what one would
expect: tar/oil on flame arrows was hidden, while drugs on needles
identified on sight.