Be more consistent in death descriptions.
The following could occur:
 "Collapsed under their own weight"
 "Succumbed to your own poison"
 "Engulfed by own cloud of flame"

The first is good, if a bit informal (see #5618).  The second isn't so
great, because the person reading the entry might not be the person who
died.  The last is abysmal.

Change the last two to use "their own" instead.
(cherry picked from commit 67487916dfbe38cfbfd8128ebb402f0448980282)