Don't suppress the display of sInv and speed of ghost_demons.
All other abilities and resistances of this kind are shown, so we should
either show them all or block them all.  Also, speed is easily noticeable
if you watch them for a couple turns.  sInv can be noticed as well if you
are invisible -- which can affect the decision whether to turn invisible
or not, but that's no different from the decision whether to use Chain
Lightning or similar highly resistible spells.

Alternatively, we could record which resistances the demon in question has
shown (and reveal speed only after a move or two), but that'd require some
actual coding effort.