Don't crash on wizmode hell exits.
If we entered Hell without ever having passed through an enter_hell
feature (i.e. if we went there with &~U), you.hell_exit is zero.
Subsequently leaving hell crashes when trying to build D:0.

Fix this in wizmode, by going to D:1 instead.  Outside of wizmode,
die with a message.

The root_branch check happens before this, so if someone wants to make a
game mode that starts in the Vestibule, that should still work.