Demonspawn mutation ordering improvements.
The ordering of the 15 (or 18 if monstrous) mutations a demonspawn gets is
now determined as follows: each mutation is assigned a random number in some
subinterval (depending on which mutation it is) of [0,1000), and then the
mutations are sorted by these numbers. Mutations that we want to be given out
late (tier 3 facets, demonic guardian 2 and 3) are given later intervals and
mutations that we want to be given out earlier are given earlier intervals.

The most noticeable effects of this change should be the following:
    * demonic guardian 2 will usually not be given until at least 1/3 of the
      way through a player's mutations;
    * demonic guardian 3 will usually not be given until at least 2/3 of the
      way through a player's mutations;
    * body-slot facets will usually finish before 2/3 of the way through a
      player's mutations.

It should also be much more adjustable than the old system.