tx-mktxt.pl: replace the -o option by -m with opposite effect.
And actually make it work (-o was ignored). I think it's easier to understand
this way.
When using -m (merge), the .ini file is loaded first, then the .txt file is
read and if some values in it are different than the .ini, they will be
updated. This mode is for updating the original english text.
When not using the option, the .txt file is generated from the .ini and anything
in it will be overwritten. Useful if there are new or deleted keys, so it's
intended for translated files.
In both mode, if the file content is unchanged, it is not overwritten with the
same content. Preserving the timestamp of the .txt will avoid useless
regeneration of the .ini by make.