Transifex sync.
Polish translation for backgrounds was tricky. Thanks for stress-testing
the workflow kilobyte :)
It got input from both git and transifex. The git updates were pushed to
transifex, but when the new translations were pulled from transifex, the
git one weren't pulled. This is because they weren't in the reviewed state
on transifex. So I had to manually "review" them on the website to have
them pulled along the new ones.
To avoid this, we need to be able to specify their mode when we push them
to transifex. If they are already in trunk, we can assume that they are
"reviewed". Transifex doesn't support the feature yet, though I've made
a FR about it.