Suppression: block evokables and item abilities
For the sake of simplicity, we have to draw a line somewhere. Our threshold is
"if it's equipped, all magical effects are blocked; if it's not equipped, it is
unaffected". Also, again for the sake of simplicity, our only exception to
"magical effects" is evoking a reaching weapon. We already suppress passive
properties of items (resists, gourmand, brands, etc.), and this commit
suppresses active abilities from items. Specifically, any item that grants an
ability in the 'a' screen no longer grants that ability while in the
suppression field, and any item that requires that it be equipped to evoke it
(rods, staves, decks) cannot be evoked in the field.