Longer descs for SoH flavours.
Eronarn proposes changing it to some weird kind of snake, but I personally
think the existing theme, of a dragon, is better.  In fact, Slavic dragons
are more of the serpent kind, to the point of having different words for
these and for Western style dragons.

Also, Chinese dragons or, say, Tolkien's description of Glaurung, are more
serpenty as well.