Auto-ID "rMut.
Only when you can be certain of it:  when it causes you to get the "You
feel odd for a moment" message, and you don't have the Mutation
Resistance mutation.  In particular, this means quaffing a potion of
mutation will not identify the amulet, as it does not give a message
when resisted, and there might have been other reasons (physiology
conflict, etc) the mutation did not take hold.