Don't suppress the Mace of Brilliance's aura.
suppressed() calls _update_agrid(), which calls halo_radius2(),
which indirectly called suppressed().  Reading a scroll of silence
with nothing else around would crash the game.

Not to say that it shouldn't be suppressed, but it will be tricky
to do so properly (without crashing and without using a half-updated
agrid).