Track whether a monster has shown it has a ranged attack.
This fixes information leaks (the other option would be to include
bows/crossbows slung across the shoulder in "comes into view" messages).

This also allows resting when there's a sentient monster on the other
side of water and it has no means of attacking us. If this proved to be
false and it did have darts hidden in a pocket but never chose to use
them before, the attack will interrupt rest as usual.

Monsters are considered to have a ranged attack they're:
* non-ghost/demon with ranged/summoning spells
  There is no randomization that would change this.
* ghost/demons who have cast _any_ spell.  It's prudent to assume that if
  they cast spells at all, they likely have something nasty.
* are of a type which usually starts with a ranged weapon (centaurs, etc),
  and are not a known shapeshifter
* were seen wielding a bow/etc
* were seen shooting/throwing something
* were seen picking a bow/dart/hand axe/...
* were seen zapping a wand