Fix a redefined monster name/desc discrepancy.
Descs speak about a tengu, orc with a torch, dwarf with a pickaxe and a
mermaid, respectively, while the statues' names say "elemental".  At least
within our meaning of that word it's kind of impossible to make a statue of
those, as elementals are shapeless lumps of the relevant material (as
opposed to Paracelsian elementals, ie sylphs, salamanders, gnomes and
undines).

I thus renamed the statues to "foo elementalist statue", although I'm not
satisfied with that either.  Any better ideas?