db_lint: allow checking languages other than English, or a subset of files.
I guess the parsing for whether a file is selected for checking is a wee bit
of an overkill.  It allows full filenames, base part ("unrand"), dir part
("descript", "database"), dir/base ("descript/unrand").

For languages other than English, a description file totally missing is not
considered an error (even if listed for checking), but if it exists at all,
missing descriptions are errors (currently for features, unrand, cards,
monsters).