Increase absdepth of Swamp/Snake/Shoals/Spider by 1, adjust unique generation there.
This is the old average absdepth; previously it was one lower because of
rounding down for both Lair and the S branches, so this change gives a better
approximation to the old state of matters.

Also, Boris, Mennas, and others were sad at not getting to kill people on
Swamp/Snake/Shoals/Spider:5 any more; now they can.