Don't move a trampler into tramplee's space until after the given attack.
This fixes crashes if either was moved off-level.  It does _not_ fix a
random monster with the same mindex continuing wuth the next attack on the
level you dropped to -- combat code is sprinkled with
attacked/defender->alive() which happily uses a stale pointer which, while
currently always pointing to an actor struct (you or menv[x]), may refer to
another monster which can happen to be alive and in range.

Also, for some reason attacker following doesn't obey traps.  This is more
fixable than it used to be, though.
(cherry picked from commit 17597ac35bb9753363a5bcb839b87c47ca043a1f)